Changes within figure(3)
Changes within figure(3) lines 190- specifically 211 and 216, I wrote simple comments next to them. There are only issues running S.dmaxset but not S.zmaxset

Lines 185-195 is the addition of code from figure(8) from the test_SWSvsFMvsBH, this is only the code for the scatter plot. I am struggling using S.SHmax since it is not in 2D.

Sorry for so many questions